fix(e2e): correct relative imports after utils folder split
Folder moves left one-level-too-shallow paths to errors, api-endpoints,
support, and locale modules, which broke type-aware oxlint in CI. Use
global fetch in keycloak helper now that Node 24 provides it natively.
